moved code to be more readable
