+++ src/handle-labels-request-change.js
@@ -1,0 +1,49 @@
+const octokit = require('@octokit/rest')()
+const client = octokit.authenticate({
+    type: 'token',
+    token: process.env.GITHUB_TOKEN
+})
+
+async function handleLabelsRequestChange(context) {
+    
+    const { repository, issue } = context.payload;
+
+    const result = await octokit.issues.listEventsForRepo({ owner: issue.user.login, repo: repository.name })
+
+    // get all latest events with the same timestamp
+    const latestEvent = result.data[0];
+    const sameTimestampEvents = result.data.reduce((acc, curr) => {
+      if (curr.created_at === latestEvent.created_at && curr.event === 'unlabeled')
+        acc.push(curr)
+
+      return acc
+    }, [])
+
+    for (const event of sameTimestampEvents) {
+      if (event.label.name === 'Release QA' && event.issue.labels.find(label => label.name === 'Failed Release QA')) {
+        // Replace Priority Label
+        const replacements = event.issue.labels.map(label => {
+          if (label.name.startsWith('Priority')) {
+            return 'Priority: 1-Critical'
+           } else {
+            return label.name
+          }
+        })
+
+        const replaceLabelResult = await octokit.issues.replaceLabels({ owner: issue.user.login, repo: repository.name, number: issue.number, labels: replacements })
+
+        // Add Remediation Label 
+        const addLabelResult = await octokit.issues.addLabels({ owner: issue.user.login, repo: repository.name, number: issue.number, labels: ['Remediation', 'Release Branch'] })
+
+        // Remove Failed Release QA Label
+        const removeLabelResult = await octokit.issues.removeLabel({owner: issue.user.login , repo: repository.name, number: issue.number , name: ['Failed Release QA']})
+
+        // Open ticket if closed
+        if (event.issue.state === "closed") {
+          const changeIssueState = await octokit.issues.update({ owner: issue.user.login, repo: repository.name, number: issue.number, state: 'open'})
+        }
+      }
+    }
+}
+
+module.exports = handleLabelsRequestChange
